+++ apps/web/src/app/admin-console/organizations/reporting/reports-home.component.ts
@@ -5,7 +5,7 @@ import { filter, map, Observable, startWith } from "rxjs";
 import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { StateService } from "@bitwarden/common/platform/abstractions/state.service";
 
-import { ReportVariant, reports, ReportType, ReportEntry } from "../../../reports";
+import { ReportVariant, reports, ReportType, ReportEntry } from "../../../tools/reports";
 
 @Component({
   selector: "app-org-reports-home",
