+++ apps/web/src/app/tools/reports/icons/report-exposed-passwords.icon.ts
