+++ apps/web/src/app/tools/reports/icons/report-inactive-two-factor.icon.ts
