+++ apps/web/src/app/tools/reports/icons/report-weak-passwords.icon.ts
