+++ apps/web/src/app/tools/reports/pages/exposed-passwords-report.component.ts
