+++ apps/web/src/app/tools/reports/pages/weak-passwords-report.component.html
