+++ apps/web/src/app/tools/reports/shared/models/report-entry.ts
