+++ apps/web/src/app/tools/reports/shared/models/report-variant.ts
