+++ apps/web/src/app/tools/reports/shared/report-card/report-card.component.html
