+++ apps/web/src/app/tools/reports/shared/report-card/report-card.stories.ts
@@ -5,8 +5,8 @@ import { Meta, Story, applicationConfig, moduleMetadata } from "@storybook/angul
 import { JslibModule } from "@bitwarden/angular/jslib.module";
 import { BadgeModule, IconModule } from "@bitwarden/components";
 
-import { PreloadedEnglishI18nModule } from "../../../core/tests";
-import { PremiumBadgeComponent } from "../../../vault/components/premium-badge.component";
+import { PreloadedEnglishI18nModule } from "../../../../core/tests";
+import { PremiumBadgeComponent } from "../../../../vault/components/premium-badge.component";
 import { ReportVariant } from "../models/report-variant";
 
 import { ReportCardComponent } from "./report-card.component";
